requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
django-polymorphic>=3.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-extensions>=3.2.3
django-unfold>=0.30.0
